feat(comparator): extract splitter comparator as a dumb component
