+++ src/components/slider/splitter-element.vue
@@ -1,0 +1,102 @@
+<script setup lang="ts">
+import { onMounted, onUnmounted, ref, Ref } from 'vue'
+import { useTranslation } from 'i18next-vue'
+import { Layer } from '@/stores/map.store.model'
+import { computed } from 'vue'
+
+const DEFAULT_STEP_ONKEYDOWN = 30
+
+const props = defineProps<{
+  sliderActive: boolean
+  sliderRatio: number
+  sliderTopLayer: Layer
+  sliderOffset: number
+  containerOffset: number
+}>()
+const emit = defineEmits(['moveSplitBar', 'escSplitBar'])
+const { t } = useTranslation()
+const sliderElement: Ref<HTMLElement | null> = ref(null)
+const styleObject = computed(() => ({ left: `${props.sliderOffset}px` }))
+
+let isDragging = false
+
+defineExpose({
+  sliderElement: sliderElement,
+})
+
+function moveSplitBar(offsetLeft: number) {
+  emit('moveSplitBar', offsetLeft + sliderElement.value!.offsetWidth / 2)
+}
+
+function onMouseDown() {
+  isDragging = true
+
+  document.addEventListener('mousemove', onMouseMove)
+  document.addEventListener('mouseup', onMouseUp)
+}
+
+function onMouseMove(payload: MouseEvent) {
+  if (!isDragging) {
+    return
+  }
+
+  moveSplitBar(payload.clientX - props.containerOffset)
+}
+
+function onMouseUp() {
+  isDragging = false
+
+  document.removeEventListener('mousemove', onMouseMove)
+  document.removeEventListener('mouseup', onMouseUp)
+}
+
+function onKeyDownRight() {
+  moveSplitBar(sliderElement.value!.offsetLeft + DEFAULT_STEP_ONKEYDOWN)
+}
+
+function onKeyDownLeft() {
+  moveSplitBar(sliderElement.value!.offsetLeft - DEFAULT_STEP_ONKEYDOWN)
+}
+
+function onKeyDownEsc() {
+  emit('escSplitBar')
+}
+
+onMounted(() => {
+  sliderElement.value?.focus({ focusVisible: true } as any) // TODO: FocusOptions
+})
+
+onUnmounted(() => {
+  document.removeEventListener('mousemove', onMouseMove)
+  document.removeEventListener('mouseup', onMouseUp)
+})
+</script>
+
+<template>
+  <button
+    ref="sliderElement"
+    @mousedown="onMouseDown"
+    @mousemove="onMouseMove"
+    @mouseup="onMouseUp"
+    @keydown.space="onKeyDownRight"
+    @keydown.right="onKeyDownRight"
+    @keydown.left="onKeyDownLeft"
+    @keydown.delete="onKeyDownLeft"
+    @keydown.esc="onKeyDownEsc"
+    class="left-[20px] absolute h-full w-[32px] block"
+    :style="styleObject"
+    role="seperator"
+    aria-controls="map-container"
+    data-cy="sliderElement"
+  >
+    <span class="lux-slider-line"></span>
+    <span class="lux-slider-arrows">
+      <span></span>
+      <span></span>
+    </span>
+    <span class="lux-slider-layer-label" v-if="sliderTopLayer">
+      <i class="fa fa-arrow-left mr-2"></i>
+      <span>{{ t(sliderTopLayer.name) }}</span>
+    </span>
+  </button>
+</template>
